Use CASPER_BACKPORT env variable for backporting
If CASPER_BACKPORT is in the environment with non-zero length, then force block
reuse in reuse_block and do NOT preload CASPER libraries.  This prevents
problems when saving libraries in older Simulink formats (aka "backporting"),
but should NOT be used for normal development.

One problem prevented is the attempted replacement of blocks from their new
name to their old name.  This would normally result in the existing block being
deleted and then readded using the old (former) name, but the deletion fails
during the "save as old version" operation because it happens in a callback
function where deleting blocks is not allowed.  The resulting exception would
disrupt the normal flow of the "save as old version" operation and leave some
things in a bad state (e.g. duplicate blocks with same name).

+++ casper_library/reuse_block.m
@@ -87,10 +87,18 @@ else,
     clog(msg, {'reuse_block_debug', 'reuse_block_reuse'});
     set_param([blk,'/',name], varargin{:});
   else,
-    msg = sprintf('%s is a "%s" but want "%s" so replacing', name, source, refblk);
-    clog(msg, {'reuse_block_debug', 'reuse_block_replace'});
-    delete_block([blk,'/',name]);
-    add_block(refblk, [blk,'/',name], 'Name', name, varargin{:});
+    if evalin('base','exist(''casper_force_reuse_block'')') ...
+    && evalin('base','casper_force_reuse_block')
+      msg = sprintf('%s is a "%s" and want "%s" but reuse is being forced', name, source, refblk);
+      % Log as reuse_block_replace even though reuse is being forced
+      clog(msg, {'reuse_block_debug', 'reuse_block_replace'});
+      set_param([blk,'/',name], varargin{:});
+    else
+      msg = sprintf('%s is a "%s" but want "%s" so replacing', name, source, refblk);
+      clog(msg, {'reuse_block_debug', 'reuse_block_replace'});
+      delete_block([blk,'/',name]);
+      add_block(refblk, [blk,'/',name], 'Name', name, varargin{:});
+    end
   end
 end % if isempty(existing_blk)
 catch ex
